refactor customer into model
