+++ src/main/java/com/codecool/shop/model/customer/Customer.java
@@ -1,4 +1,4 @@
-package com.codecool.shop.customer;
+package com.codecool.shop.model.customer;
 
 public class Customer {
     private String name;
